Fix unread count query datatype mismatch
Convert feed.ID to int64 in the Datastore query filter to match the FeedID
field type in ArticleEntity. This fixes the issue where articles exist with
correct feed IDs but the query wasn't finding them due to type mismatch.

+++ internal/database/datastore.go
@@ -694,7 +694,7 @@ func (db *DatastoreDB) GetUserUnreadCounts(userID int) (map[int]int, error) {
 	// For each feed, count unread articles
 	for _, feed := range userFeeds {
 		var articles []*Article
-		q := datastore.NewQuery("Article").FilterField("FeedID", "=", feed.ID)
+		q := datastore.NewQuery("Article").FilterField("FeedID", "=", int64(feed.ID))
 		_, err := db.client.GetAll(ctx, q, &articles)
 		if err != nil {
 			return nil, err
